Lint cfg

+++ euler/3.cpp
@@ -5,7 +5,7 @@
 // The prime factors of 13195 are 5, 7, 13 and 29.
 // What is the largest prime factor of the number 600851475143 ?
 
-using ll = long long;
+using ll = int64_t;
 
 
 bool is_prime(ll k) {
